--- a/frontend/components/mention/MentionsSection.tsx
+++ b/frontend/components/mention/MentionsSection.tsx
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
-// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2023 - 2024 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -53,7 +53,6 @@ export default function MentionsSection({title, mentions}: MentionsSectionProps)
               <MentionViewList
                 key={key}
                 title={title}
-                type={key}
                 items={items}
               />
             )